+++ tests/testthat/test-joins.r
@@ -198,6 +198,33 @@ test_that("disallow empty string in both sides of suffix argument (#2228)", {
   )
 })
 
+g <- data.frame(A = 1, A.x = 2)
+h <- data.frame(B = 3, A.x = 4, A = 5)
+
+test_that("can handle 'by' columns with suffix (#3266)", {
+  j1 <- inner_join(g, h, "A.x")
+  j2 <- left_join(g, h, "A.x")
+  j3 <- right_join(g, h, "A.x")
+  j4 <- full_join(g, h, "A.x")
+
+  expect_named(j1, c("A.x.x", "A.x", "B", "A.y"))
+  expect_named(j2, c("A.x.x", "A.x", "B", "A.y"))
+  expect_named(j3, c("A.x.x", "A.x", "B", "A.y"))
+  expect_named(j4, c("A.x.x", "A.x", "B", "A.y"))
+})
+
+test_that("can handle 'by' columns with suffix, reverse (#3266)", {
+  j1 <- inner_join(h, g, "A.x")
+  j2 <- left_join(h, g, "A.x")
+  j3 <- right_join(h, g, "A.x")
+  j4 <- full_join(h, g, "A.x")
+
+  expect_named(j1, c("B", "A.x", "A.x.x", "A.y"))
+  expect_named(j2, c("B", "A.x", "A.x.x", "A.y"))
+  expect_named(j3, c("B", "A.x", "A.x.x", "A.y"))
+  expect_named(j4, c("B", "A.x", "A.x.x", "A.y"))
+})
+
 test_that("check suffix input", {
   expect_error(
     inner_join(e, f, "x", suffix = letters[1:3]),
@@ -215,6 +242,9 @@ test_that("check suffix input", {
     fixed = TRUE
   )
 })
+
+
+# Misc --------------------------------------------------------------------
 
 test_that("inner_join does not segfault on NA in factors (#306)", {
   a <- data.frame(x = c("p", "q", NA), y = c(1, 2, 3), stringsAsFactors = TRUE)
@@ -319,7 +349,7 @@ test_that("left_join by different variable names (#617)", {
   expect_equal(res$y2, c("foo", "bar", "foo"))
 })
 
-test_that("joins support comple vectors", {
+test_that("joins support complex vectors", {
   a <- data.frame(x = c(1, 1, 2, 3) * 1i, y = 1:4)
   b <- data.frame(x = c(1, 2, 2, 4) * 1i, z = 1:4)
   j <- inner_join(a, b, "x")
@@ -721,6 +751,9 @@ test_that("inner join not crashing (#1559)", {
   for (i in 2:100) expect_equal(res[, 1], res[, i])
 })
 
+
+# Encoding ----------------------------------------------------------------
+
 test_that("join handles mix of encodings in data (#1885, #2118, #2271)", {
   with_non_utf8_encoding({
     special <- get_native_lang_string()
@@ -802,6 +835,8 @@ test_that("left_join handles mix of encodings in column names (#1571)", {
     expect_equal(res[[special]], 1:6)
   })
 })
+
+# Misc --------------------------------------------------------------------
 
 test_that("NAs match in joins only with na_matches = 'na' (#2033)", {
   df1 <- data_frame(a = NA)
@@ -921,5 +956,91 @@ test_that("join handles raw vectors", {
   expect_identical(
     inner_join(df1, df2, by = "r"),
     data_frame(r = as.raw(3:4), x = c(3:4), y = c(3:4))
+  )
+})
+
+test_that("joins reject data frames with duplicate columns (#3243)", {
+  df1 <- data.frame(x1 = 1:3, x2 = 1:3, y = 1:3)
+  names(df1)[1:2] <- "x"
+  df2 <- data.frame(x = 2:4, y = 2:4)
+
+  expect_error(
+    left_join(df1, df2, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    left_join(df2, df1, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    right_join(df1, df2, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    right_join(df2, df1, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    inner_join(df1, df2, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    inner_join(df2, df1, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    full_join(df1, df2, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    full_join(df2, df1, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    semi_join(df1, df2, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  # FIXME: Compatibility, should throw an error eventually
+  expect_warning(
+    expect_equal(
+      semi_join(df2, df1, by = c("x", "y")),
+      data.frame(x = 2:3, y = 2:3)
+    ),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  expect_error(
+    anti_join(df1, df2, by = c("x", "y")),
+    "Column `x` must have a unique name",
+    fixed = TRUE
+  )
+
+  # FIXME: Compatibility, should throw an error eventually
+  expect_warning(
+    expect_equal(
+      anti_join(df2, df1, by = c("x", "y")),
+      data.frame(x = 4L, y = 4L)
+    ),
+    "Column `x` must have a unique name",
+    fixed = TRUE
   )
 })