+++ NEWS.md
@@ -45,6 +45,8 @@
   input in mutating operations and `mutate(df, "foo")` creates a new column by
   recycling "foo" to the number of rows.
 
+* Fixed rare column name clash in joins with non-join columns of the same name in both tables (#3266).
+
 * `select()` and `vars()` now treat `NULL` as empty inputs (#3023).
 
 * Add error for `distinct()` if any of the selected columns are of type `list` (#3088, @foo-bar-baz-qux).
@@ -53,9 +55,13 @@
 
 * Better error message if dbplyr is not installed when accessing database backends (#3225).
 
+* Corrected error message when calling `cbind()` with an object of wrong length (#3085).
+
 *  Fix `row_number()` and `ntile()` ordering to use the locale-dependent ordering functions in R when dealing with character vectors, rather than always using the C-locale ordering function in C (#2792, @foo-bar-baz-qux).
 
 *  `distinct()` now supports renaming columns (#3234).
+
+* Better error message when joining data frames with duplicate column names. Joining such data frames with a semi- or anti-join now gives a warning, which may be converted to an error in future versions (#3243).
 
 # dplyr 0.7.4